Move file and BIO reading out to ssl_misc.c
